+++ macOS/framework/SKTiled.h
@@ -12,4 +12,4 @@
 FOUNDATION_EXPORT double SKTiledVersionNumber;
 FOUNDATION_EXPORT const unsigned char SKTiledVersionString[];
 
-
+// In this header, you should import all the public headers of your framework using statements like #import <SKTiled/PublicHeader.h>